+++ src/main/java/wdsjk/project/avitobalancemicroservice/dto/request/DepositRequest.java
@@ -2,6 +2,8 @@ package wdsjk.project.avitobalancemicroservice.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 
+import java.math.BigDecimal;
+
 public record DepositRequest(@NotBlank(message = "Can't be blank!") String userId,
-                             double amountOfMoney) {
+                             BigDecimal amountOfMoney) {
 }
